Display helpful HTML attachment error on edition summary page
We had to fix a few tests that were passing when they shouldn't have been. In particular, it wasn't possible to create a new draft of a document with an invalid HTML attachment. That has now been fixed by skipping validation during new draft creation, which we think is a safe choice.

We added a more human-friendly label for the HTML attachments edition attribute and improved the message added by the govspeak contact embedding validator.

We had to forcibly disable association validation for attachments and html_attachments on the attachable trait. This was because, by default, when an association is loaded and a custom validation context is in use (in our case :publish), Rails will validate the models in the association, which is somewhat unexpected behaviour. This resulted in generic messages such as "attachments is invalid" being displayed to users, which were not very useful. See https://github.com/rails/rails/issues/45209 for more information.