Remove unnecessary wrapper class
In every case where we call `bare_govspeak_to_html`, we seem to
wrap it in the `<div class="govspeak">` element. So let's just
make that part of the output of `bare_govspeak_to_html`.

Note that `tag` isn't available in all of the contexts in which
this is called, so I've gone for simple interpolation instead.

It was added in c9b34539a9af63e51c5748df0b1d4c5d9305c3f0 and
appears to have been added for easier testing rather than because
it was required.

And in any case, we're hoping this is a stepping stone towards
removing GovSpeak rendering from Whitehall altogether (it should
happen in Publishing API instead). So this `<div>` wrapper should
eventually go away.