Make `add_manual_heading_numbers` a pre-processing method
We originally semi-supported adding a span around "0GD"-like headings for
HMRC, in https://github.com/alphagov/whitehall/commit/07ce5eac1827397236a0cf20074ba95c5663b5e0.
But it doesn't seem this was asked for - the issue was that a space was being
added between the "0" and the "GD", and including it in the manual numbering
code seemed to be the best way to solve it at the time.
We now feel the best thing is to not treat such headings as manually numbered.