Enable backdating of first publication date for standard edition types
We change the behaviour of the first published at component so that it renders a hideen field with the value set to false for the first edition of edition types that have backdating disabled, instead of rendering nothing at all. This saves us from having to override the `previously_published` value for editions where backdating is disabled.