Remove `add_manual_heading_numbers` logic
See thread at
https://github.com/alphagov/whitehall/pull/10589#discussion_r2300382310

All this logic did was wrap the 'manual number' in a span. Example:

Govspeak:                    `## 1. Heading`
Rendered HTML with the rule: `<h2 id="heading"><span class="number">1. </span>Heading</h2>`
Without the rule:            `<h2 id="heading">1. Heading</h2>`

The addition of the `<span>` doesn't actually change any of the
visual or screen-reader experience on the frontend, so it's not
clear why this is needed anymore. Deleting.