Persist image selection on new Standard Edition drafts
Re-drafting a published edition causes the images to be recreated anew. This meant that the "sidebar image" selection for published History Pages (or, more generally, block image selections for Standard Editions), did not persist when creating a new edition.

Nonetheless, images do share an ImageData across editions. This ImageData cannot be replaced (unlike for Attachments), so it is safe to store the ID of the ImageData as a reference in the `block_content` field (the field on the edition model which stores the content corresponding to the schema). We can thus use it as the means to set the selection in the (Sidebar) Image drop-down.