Remove unnecessary validation rule around 'internal' links
If publishers choose to link to the user-facing slug
(`/government/foo/bar`) instead of the Whitehall admin slug for
that page (e.g. `/government/admin/foo/123`), we've traditionally
flagged that as a validation error and prevented them from doing
so.

[Asking around the content community](https://gds.slack.com/archives/CACV3TACU/p1752572986431479),
there is general consensus that, whilst linking to the 'admin'
path can have its uses, it shouldn't be enforced at the
validation level. Softening this rule should reduce friction for
our publishers and also reduce the number of editions considered
'invalid' in our system health reporting.

The only benefit to using the 'admin' links is to batch up
publishing of several related drafts. Testing on integration:

- Draft Edition A, internal-admin-linking to Draft Edition B
- Draft Edition B, internal-admin-linking to Draft Edition A
- No links visible on preview (known bug), but…
- …set them both to schedule-publish at the same time, and when
  they published, they both had correct links to the published editions.

But there's no reason to force publishers to link to the 'admin'
version of a thing if that thing is already published! It only
adds complexity for both the publisher and the underlying code.

JIRA: https://gov-uk.atlassian.net/browse/WHIT-2350