Add migration to remove 'visual_editor' boolean from tables
These fields are no longer used as of #10603.

The `ignored_columns` addition to Edition and Attachment can be
removed after the migration has run. It is enforced by the
strong_migrations gem to prevent app code from referring to the
DB fields (which could happen from a worker or something, in
theory, even though we've already removed the code).