Notify Slack if CI fails on the 'main' branch
We sometimes have cases where CI passed on the build of the branch
in the PR, but subsequently fails when the PR has been merged (e.g.
due to a flaky test). In such cases, the deployment is blocked and
so it is easy to assume that your merged code has been deployed
when in fact it hasn't! This [happened to us recently](https://gds.slack.com/archives/C013F737737/p1755507962456949).

Luckily for us, Platform Engineering have already built a
workflow to notify us of this very situation. They've documented
how to use the workflow, here:
https://github.com/alphagov/govuk-infrastructure/blob/main/.github/actions/report-run-failure/README.md

This commit follows those instructions, with the small tweak of
specifying the name of the `action.yml` file, which was being
enforced as good practice directly in my VS Code editor.