Fix error when generated notice becomes too large
`republish_documents_by_content_ids` can generate a notice that will
cause an error to occur if the list of `content_ids` provided is large
enough. This is because the session generated as a result of the
notice exceeds the size limit due to the notice including the list of
all the `content_ids`.

To overcome this, if the `content_ids` is over 20 then the notice will
not include the `content_ids`.