Prevent deletion of assets associated with unpublished documents
Previously, when discarding a new edition of an unpublished document, any assets would get deleted. That meant that if a new edition of that document was subsequently published, the assets would return a 404 response in Asset Manager.

These changes ensure that we only delete the assets, in the case of an edition discard, if they have never been associated with a published edition.

Checking the attachments number is suitable for this purpose, because:
 - in the case of a first draft edition there would only be one attachment for the `AttachmentData`
- for editions that have already been live, the attachment gets made anew when re-drafting the edition, whilst sharing the `AttachmentData` with previous attachments, meaning the size of the attachments will be > 1.