Save `crop_data` to `ImageData`
Allows images that are too big to be saved. When this happens
the user will be asked to crop the image on the `index` page. This
change means that images can be recropped, the cropping data can be
used in other contexts and multiple images can be uploaded without
validation errors if all/some of them are too big.