Refactor edition organisations association trait module
This is a pure readability improvement. The previous implementation of this module was hard to follow and to reason about.

It's worth noting why we don't use normal associations for lead and supporting organisations. Rails persists association changes to has_many associations as soon as they're assigned, meaning that validation never runs. This meant when using real associations, users experienced a server error instead of a validation error. We can get around this by implementing the faux associations ourselves.

We also remove some methods that are no longer used.