Rename GovspeakLinkValidator to InternalPathLinksValidator
All that remains of the original GovspeakLinkValidator is a check
that links to any internal paths (e.g. `/government/news/123456`)
are correctly formatted (i.e. start with a leading `/`). These
'absolute path' links aren't yet sent to Link Checker API for
checking, and the potential for accidentally flagging perfectly
valid relative links is too much of a risk to start doing this
check in Link Checker API. Moreover, the absolute path architecture
in Whitehall is quite unique: paths are swapped out for their
public-facing base path at publish time (e.g.
`/government/news/123456` -> `/government/news/this-is-my-published-story`),
which is very Whitehall specific, and so its validation should
probably continue to live in Whitehall and not in Link Checker API.