Tighten language and tests
Rather than "broken links" (which conflicts with / duplicates the
'broken links report' section below it), we now refer to these as
'invalid admin links': Whitehall-specific subtypes of link, which
is all that this validator is checking for now.

We've also added a test for the case where a publisher accidentally
prefixes an admin link with two or more `/` slashes. Previously,
the hint text would have noted that it wasn't a valid admin link
and ask them if they meant `///{inputted value}` (i.e. it would
suggest adding even more slashes!).