Add support for images
Content Publisher only produces two variants of images: 960 and
300 sized. Whitehall expects (and enforces) several other variants:
712, 630, 465 and 216. If said variants are missing, then the
edition is considered invalid, and also the images screen in the
UI is stuck showing a 'processing' message. These behaviours are
largely triggered by the various `all_asset_variants_uploaded?`
method definitions sprinkled throughout Whitehall.

My first instinct was to soften the check to only class s960 and
s300 variants as 'required': https://github.com/alphagov/whitehall/pull/10646
But this would be very specific to just this use case, and in theory
I'd want to be dynamically detecting which variants are required
depending on the context.

I then looked at dropping the check altogether: the vast majority
of the time, if the 'original' asset has been uploaded, so has its
variants, so it seemed a low risk change worth making to simplify
the code: https://github.com/alphagov/whitehall/pull/10647

But on reflection, these other variant sizes are potentially needed.
For example, when featuring a news article on an org homepage,
the s465 variant seems to be needed. So, in the unlikely event
that a publisher wants to feature a >2 year old news article on
their homepage, they wouldn't really be able to unless that
variant existed.

That left us with two remaining options:

1. Reupload the original image through Whitehall's Carrierwave
   pipeline, to have it generate all variants afresh
2. Reuse a larger variant for a smaller one (e.g. the s300 Content
   Publisher variant can be used in place of the s216 sized variant,
   given browsers typically take whatever image and just scale it
   down as appropriate.

Option 1 is more complicated to implement, and raises several
other questions:

- How do we publish the new images, given we're importing documents
  'already published' (rather than as drafts which we later publish)?
- What do we do with the old assets? Delete? Redirect? Leave alone?

Option 2 therefore feels the pragmatic way forward, given we
already have a working import solution for file attachments that
we can apply to images too.