Show collaboration with FindAndReplaceEmbedCodesService
The `PreviewController` renders a `preview` template which
calls `GovspeakHelper#govspeak_to_html`. This includes a
`preprocess_govspeak` method which collaborates with the
`ContentBlockManager::FindAndReplaceEmbedCodesService`.

Now that the ContentBlockManager is moving out of its
location as an engine within Whitehall we need to:

1. relocate this `FindAndReplaceEmbedCodesService` within
Whitehall proper before the Content Block Manager "engine"
is deleted

2. change the `FindAndReplaceEmbedCodesService` to retrieve
any referenced content blocks from the Publishing API rather
than from the local db

In this commit we start by adding a test to demonstrate how:

- the `FindAndReplaceEmbedCodesService` is given govspeak
markup containing embed codes and returns an HTML
presentation of the block

- that HTML version of the content block is included in the
rendered preview template