Add association links to standard edition presenter
We also remove the edition links from the content item details for standard editions, as Whitehall does not use these for any other document type. Whilst it would eventually be desirable to migrate to using edition links once they support nesting, we will have to undertake this as a separate initiative. Hopefully the sturcture will be the same, so the migration will not be too arduous.