Add DocumentImporter class and tests for basic properties
This class will be called by a worker in a later commit, for
doing the actual import. This first commit just sets up the
basics: slug, title, body, state, timestamps etc. In subsequent
commits we'll tackle the trickier aspects that need importing.

Note that I originally started using Timecop to overwrite the
system time at the point of import, so that documents were being
imported 'at the time' they were created. But there was a major
problem with this: Timecop pollutes time globally and is only
intended to be used in test suites. Consequences:

1. Running the bulk import (batching lots of Sidekiq workers)
   didn't work as expected, as several time overrides would happen
   in parallel and thus documents would be given unpredictable
   created_at dates.
2. During the bulk import with all the time overwriting going on,
   there was every chance that some unrelated live production
   activity could be affected and its timestamps effectively
   corrupted.

Instead, I now import the docs as-is and then updated the created_at
values wherever needed, after the fact.