Add public changenotes
In the Content Publisher export, we have the full array of public
changenotes, but we can only associate an edition with one
changenote. We've decided not to recreate the full edition history
of imported documents, instead importing 'flattened' published
editions. This is the thorniest bit of that 'flattening': we need
to consolidate all of the changenote history into one sentence.