Data migration to delete orphaned HMRC customs guidance assets
The state of these assets was modified directly in Asset Manager, so we need to reapply the Whitehall state to synchronise the two systems. The assets became orphaned when a redirect was created for them in Asset Manager, but Whitehall's "AssetManagerAttachmentRedirectUrlUpdateWorker" was retrying still and subsequently removed the redirect because there is no redirect configured for the attachment that owns these assets in Whitehall.