Avoid sending `details.headers` for unsupported content types
The top level `details.headers` property is required for History
Pages:
https://github.com/alphagov/publishing-api/blob/4005b72ce5996ade8b83eda568c38afcbe63fd79/content_schemas/formats/history.jsonnet#L19-L20

It is not allowed by the schema for News Articles, nor do we see
much purpose in allowing it:
https://gds.slack.com/archives/C02L13S214K/p1757428495759379

It is therefore a configurable property in our JSON. We can always
revisit this later (e.g. split out into separate presenter for
History pages, or adding support for headers to the News Article
jsonnet, etc).