Update `type` filter parameter to refer to document subtype
Currently, if choosing a subtype from the dropdown list, the
search filter will be set to the ID of that thing, e.g.
`news_article_2` (referring to press releases, as per the
[NewsArticleType](https://github.com/alphagov/whitehall/blob/0172090e2e522c8662d94bedcf4c4036bece03a1/app/models/news_article_type.rb#L39)
configuration).

If we don't change this, we'd need to find some way of retaining
the mapping of the ID to the subtype for both legacy and new
'StandardEdition' flavoured subtypes, which would quickly become
messy. So let's fix this first: we now just pass the name of
the subtype directly (i.e. `type=press_release`). This will make
it easier to include StandardEdition documents in the search,
since they'll share the same name in the `configurable_document_type`
property.

This commit:

- Removes the `by_type` scope (which was only used by the edition
  filter) in favour of doing it inline in the edition filter (and
  having more control over it)
- Removes references to 'subtypes', which are no longer used to
  power search.
- Dynamically looks up StandardEdition or 'legacy' subtypes by
  key, rather than ID.