Allow associations to present multiple types of links
News stories need to present all organisations in the links object and also a primary publishing organisation. This was not possible with the previous configurable association API, so we now just allow each association to present any links it needs to and merge them all together into the links hash.