Simplify inline file attachment enabling logic
For all extant Whitehall document types, if HTML attachments or external attachments are enabled, then inline attachments are not. We therefore enforce that constraint rather than relying on developers to remember to configure the two separate `allows` methods to maintain it correctly. We had failed to do this for call for evidence and consultation outcomes, which could result in unexpected behaviour if a publisher tried to embed an HTML attachment inline in the summary field. Fortunately there were no examples of this in the database.