Remove govspeak renderer and block attachment renderer
The govspeak renderer was used to build a rendering context for govspeak and attachment rendering. However, by using the Application controller renderer inside the helpers instead of the view context render method, we can leverage the existing rendering context instead, making the GovspeakRenderer redundant. We can then just include the GovspeakHelper module everywhere that we were using the renderer, as it just a plain old Ruby module.

The block_attachments method in the attachments helper was only used from Publishing API presenters, so it has been moved into a Presenter specific helper.