WIP PUT content when republishing unpublished editions
Currently, when republishing unpublished editions we only call the

    POST /v2/content/:content_id/unpublish

endpoint. This updates details of the unpublishing, but doesn't touch
fields on the edition itself (things like rendering_app, which we
occasionally care about).

For reasons I don't understand, whitehall seems to consistenly refer to
the put_content action in publishing API as "save_draft", which is
really pretty confusing given that it's often called for published and
withdrawn editions, which aren't drafts. I'm reluctant to go down the
rabbit hole of unpicking this.

The save_draft method is synchronous (i.e. it calls the API directly,
without putting a job on the queue), and I haven't wrapped it in a
worker. I figure this is okay as we're already inside the
PublishingApiUnpublishingWorker, and other things are called
synchronously in this class (including this same method in
republish_pre_publication_edition).

WIP because I haven't thought about tests at all, I'm just interested in
what CI thinks.