PUT content when republishing unpublished editions
Currently, when republishing unpublished editions we only call the

  ```
  POST /v2/content/:content_id/unpublish
  ```

endpoint. This updates details of the unpublishing, but doesn't touch
fields on the edition itself (things like `rendering_app`, which we
occasionally care about).

This means that when the presenter is updated and content republished in
bulk, any changes in that presenter do not take effect for these
unpublished content items.

Therefore adding an additional request to

  ```
  PUT /v2/content/:content_id
  ```

prior to making the POST request.

This will have the effect of creating a draft in Publishing API. By
changing `allow_draft` to true in the POST request, we will then use the
newly created draft when updating the unpublishing.

Co-authored-by: Richard Towers <richard.towers@digital.cabinet-office.gov.uk>