Support bulk republishing of documents that are Standard Edition only
The previous implementation of #republish_all_by_type attempted to constantize the content type to the method. This will raise an exception for subtypes of standard edition and meant that exclusive types such as News Story would error here. This constantisation is now only used by #republish_all_by_non_editionable_type so it has been moved to there.