Remove lead image concern from speeches
Never used. The logic to render anything on the Images tab (including defaulting behaviour and lead image guidance), is actually:
```
  def render?
    edition.can_have_custom_lead_image?
  end
```
which is only true if you also include the CustomLeadImage, which Speeches don't have.

Alternatively, we might want to actually have the same behaviour here, where we DO show the default image, and can opt out of it.