WIP - make all document behave types the same
- it makes NA behave like case studies
- case studies actually have an option to opt out of the behaviour of defaulting, so share that behaviour with NA

Nonetheless, to implement this, if we want all logic to be on the FE, we need to send some flag like `show_default_image` through.

Speeches can be made to work the same, i.e. offer the option to opt out of the defaulting behaviour, which is currently not shown at all in wh.

The concern here is that all that logic lives in FE.