Dynamically include configurable content types in Bulk Republishing options
Tests were updated to remove references to Rails.application.eager_load!. This made testing unpredictable on CI vs locally. The eager load picks up classes that are defined in tests as well as app models. It didn't seem like the increase in complexity was worth the benefit so tests were changed to reference static lists of types.