Implement presence validation for configurable documents
We want to leverage active record validations, so in order to do this we need to provide an Active Model object to represent the block content. Usually this can easily be done using active model attribute type casting.

However, we can't use Active Model attributes here because they rely on you knowing the type you want to cast to at initialization time. We're often fetching the standard edition from the database and checking its configurable document type to know what schema to pass to the block content model, so that won't work for us. We therefore use a setter and getter to serialise and deserialise the raw hash to and from a model object.

The same applies to active record validations. Generally you need to declare the validations at initialisation, but we don't have the type information to do that. We therefore use the `with_validator` instance method exposed by Active Model to apply our validations, instead of the usual validation-related class methods.

To allow the content to differ by document type, we back our block content type with a Struct which is defined dynamically based on the configuration for the document type.

We're removing the JSON schema validation immediately as the expectation is that Active Record validations will be used for all configurable document validation in future. We will also be discontinuing the use of JSON schema to define the types.