Implement presence validation for configurable documents
We want to leverage active record validations, so in order to do this we need to provide an Active Model object to represent the block content. Usually this can easily be done using active model attribute type casting, but unfortunately the Globalize gem breaks attribute type casting for translated attributes. We therefore implement our own type casting in the standard edition block content module.

To allow the content to differ by document type, we back our block content type with a Struct which is defined dynamically based on the configuration for the document type.