Add validation to ensure that alt text is more than quotes/spaces
We were seeing several features in the database where the alt text
was literally `""`, and coming through to the frontend as
`alt="\"\""`, which was a poor experience for screen readers. We
expect some publishers were misenterpreting guidance around using
empty alt attributes for decorative images.

This validation rule ensures that if a publisher only passes quotes
(or spaces) to the alt text field, then the featuring will not be
considered to be valid.

Jira: https://gov-uk.atlassian.net/browse/WHIT-2593