Create a new lead image block
- The default image value is passed to the LeadImageSelect.
- The computation of what the default lead image is will reside in the concern.

This is to best isolate the code for lead. This means we don't need to configure a setting in the json file for news story, to tell it that it can have a lead. Just using this block expressed that the image is lead, and will have the behaviour whereby it defaults to its organisation image, some custom payload etc.

No need to inherit from the regular lead image, though behaviours are shared. There is enough difference in path and payload that inheriting feels redundant.

The block had to be added to the json (`configurable-document-type.schema.json`) for validating the schemas.