Adds validation rule to ensure limit on length of body attribute
Replicating the validation rule that exists here:
https://github.com/alphagov/whitehall/blob/e88cc175b55eaf553694aefdbf18af87b6ead7e2/app/models/edition.rb#L61C76-L61C86

I had to change how we call the validator. The
`with_indifferent_access` casting did not work as expected - the
validator expects staight symbol keys. And the validator was also
stripping out the `attributes` property and failing to find the
thing to validate.