Send correct lead image and image payload
- Lead image has a custom payload, including high resolution url.
- If the custom image is not uploaded, then the org default or a placeholder are sent.
- Default news images don't have a caption at all (we never upload it). We're working straight with the `FeaturedImageData` model here, which is what a default news image is.

Also fixed the image payload -> as per the previous `NewsArticlePresenter` behaviour, given that the caption may also be blank (""), we used to convert it to nil when forming the payload. Since the caption is not required in the schema, I changed it so that we don't send the property at all if it's nil.