Add safe HTML validator to body
Unlike the other custom validators, we're not adding
`attribute: :body` in edition.rb for this one. We want the
validator to apply to ALL attributes, otherwise the
`should_protect_against_xss_and_content_attacks_on` check
fails for a large number of content types.

SafeHtmlValidator has also been updated to normalise its configured attribute
names and change keys to symbols, making it indifferent to whether models use
string or symbol keys. This removes the need for BlockContent to transform its
keys and decouples the two classes cleanly.

It also checks whether `changes` and `marked_for_destruction?` are
defined, so that the validator can run in different contexts.