Add new validation rules to History pages
This commit adds double ledger testing as per
https://docs.publishing.service.gov.uk/repos/whitehall/adr/0006-config-driven-content-types.html#3-principles

I had to add a `bypass_cache` keyword arg to the `find` method of
`ConfigurableDocumentType`, as our tests clobber the value of
`@types`. In the case of these new double ledger tests, we do
want to test the REAL definitions of the content types, not a
mock value.