Check that organisation methods are available
Just like we check for worldwide organisations, we should check for organisations and lead organisations. It could be that the associations have been missed out from whatever type of edition (such as `StandardEdition`) is trying to implement the lead image.

Alternatively, we could include the Organisations concern here, but that's included as an association, conditionally, so including it here would make that conditional inclusion redundant.