Allow non-www variations of allowlist hosts
We had a publisher try to redirect to the non-www version of a
host on the allowlist. That should be allowed.

Note: it's tempting to try to "refactor" the validator to remove
the "." prefix on the hosts and therefore simplify the host
checks - but the "." is important to keep as otherwise a publisher
could redirect to a domain that _isn't_ on the allowlist, e.g.
`examplepolice.uk` vs `police.uk`.

Jira: WHIT-2611
Zendesk: 6251559