Introduce type casting for block content attributes
This allows us to persist values using native JSON types rather than everything being a string. This is useful as it saves block objects such as the image select block from having to do the type coercion. It also means that we can properly represent a nil value, which saves us from having to actively omit empty string values from the content in cases where they aren't wanted, such as in the publishing API payload.

The basic implementation is fairly crude but we can add sophistication as our needs develop.