Make configurable document type config access thread safe
This is more a vanity commit for technical correctness than anything else, as the operation to load the config files will be idempotent (as long as the config files don't change on disk which... they really shouldn't). But I think it's good hygiene and might save a bit of time on someone's request if that operation gets slow due to lots of content types.