Introduce worldwide organisation concern to standard edition
Included concern in standard edition model. Also made the validation style consistent with the approach in the organisations concern - dropped the `skip_`- like methods for validation, in favour of using "_required?".

The worldwide organisation association is not always required (i.e. validated for presence). For example, case studies, once migrated to the config-driven implementation, will "enable" the association, but not require it. In order to address this, we've introduced another key, "required", in the associations array. This is only taken into consideration for associations that have a `_required?` method defined.

Other associations, like the Organisations association, for example, would ignore the "required" key completely, as they are always required. This might change in the future, for consistency.

The "required" behaviour will be introduced for world locations in upcoming commit.