Send placeholder url in the payload
This will allow us to remove surprise code from frontend, where this logic is duplicated.

We use the placeholder to render it wh-side anyway. If we maintain it in two different places, we risk FE changing it out of sync with WH, and showing the publisher an image that might not appear on the rendered page.