Stop retaining 'invisible taxons' [WHIT-2627]
This commit removes all logic that attempts to 'hold onto' taxons
that had been previously tagged to, but are no longer valid (i.e.
they've been deleted or made draft).

In #3564, it looks as though there was a deliberate decision to
retain these 'invisible' taxons - but the PR and commits do not
explain the reason _why_ this was considered a problem:

> Fixes the problem that invisible taxons that were tagged to
> content (eg from legacy taxonomies) would be overwritten when a
> publisher edits the taxon taggings for a piece of content

What we _do_ know is that the reverse is true: retaining 'phantom'
taxons _causes_ problems, such as previews of documents now
showing the wrong ('phantom') taxons in breadcrumbs etc, because
they’re associated with draft taxons. When publishing the document,
those draft taxons are (correctly) not rendered, causing a
confusing experience for publishers.

I did wonder if there's an existing workflow whereby publishers
want to tag their content to a draft taxonomy ahead of time, but
publishers can't actually select draft taxons to tag (in the
Whitehall UI), so this removal commit shouldn't be breaking any
workflows.