Use `fieldset` within `datetime_fields`
The `datetime_fields` component was consistently being used within a
`fieldset` component. Therefore it made sense to move `fieldset` into
the `datetime_fields` component. If `fieldset_legend` is specified, then
the component will be rendered within the `fieldset` component.

This has an added benefit of being able to use `legend` within the
`datetime_fields` to generate tracking attributes. The text of `legend`
is used to construct `section` parameters, having this within the
template itself means the tracking code can be simplified.