Add support for history pages in Whitehall search
I have added configurable_document_type_options which currently only returns history pages, as it is the only config-driven document type that isn't a news article sub-type. I have also created a method that merges the two types of legacy and config-driven documents into an array, and using this combined list instead of the original type_options_container so that it is displayed in the dropdown.