Fix presentation of international delegations
International Delegations are unusual because the World Location News is
presented at the base path (and with the Content ID) of their World
Location [1].

At the moment, we have two presenters both publishing at this path:
`WorldLocationPresenter` and `WorldLocationNewsPresenter`.

This means that if the `WorldLocationPresenter` was the most recent
presenter to publish at this path, the International Delegation page has
no rendering app and therefore returns a 404 on the live website.

Updating the `WorldLocation` model so it is never published when the
object is an International Delegation.

1: https://github.com/alphagov/whitehall/blob/18a7f515af99a8ccb6274a3460ee9e7c971f56d0/app/presenters/publishing_api/world_location_news_presenter.rb#L68-L70