Deep merge (rather than overwrite) `block_content` on update
We're going to split updating a StandardEdition's `block_content`
across multiple screens, rather than it all being done on the
one `Document` screen/tab, as we're intending to move the "Lead
image" selection to the `Images` tab. We therefore need to edit
the update behaviour to only update the `block_content` child
properties that have been passed - not overwrite the entire
object - so that partial updates can be applied via separate
workflows.

This commit overrides the `block_content=` behaviour (which is
also used under the hood by `update` and `update!`) so that JSON
is merged, rather than overwritten. Some things to note:

1. I've not added a test for 'deep merging', as we don't yet have
   any properties that are more than one level deep.
2. I've tested that passing an unrecognised property is silently
   dropped on-merge. I could have raised an exception instead, but
   imagine we should have that error handling covered at a higher
   level.
3. I've tested that if there is an unrecognised property already
   saved in an Edition's JSON, that it is silently dropped,
   effectively 'normalising' an edition on each save. I would hope
   we're never in a position where we encounter that behaviour,
   as if we ever drop support for a property such that it would
   become unrecognised, we also ought to have applied a data
   migration to drop the property on existing content.
4. I've added a new test file, rather than add to the existing
   `standard_edition_test.rb` file, which otherwise risks growing
   to an unmanagable size (and is in any case more geared towards
   testing the behaviour of individual properties).

Making these changes at the model level rather than the controller
level means that this behaviour automatically applies to both the
EditionsController and the StandardEditionTranslationsController.