Redirect back to Images tab when updating JSON from that tab
The default behaviour was that it would redirect back to the
Document tab, which could be a little surprising for the publisher.

I've had to add the behaviour to the Edition controller, as that
is where the redirect happens inline. But I've had to add the test
to the StandardEdition controller, as `patch :update` is not a
route on Edition (it's a polymorphic route that gets routed to
a concrete controller, e.g. PublicationsController, or in this
case, StandardEditionsController).