Add new 'configurable_document_group' property
We need to 'group' some content types together, for the purposes
of Whitehall UI and search capability, e.g. an interstitial page
for choosing a sub-type of 'news article' when creating a new
document, or for Whitehall to know that filtering by 'news_article'
should bring up corresponding documents for news stories, press
releases and so on.

We _could_ use `publishing_api_schema_name` for this, but that
would constrain us if we wanted to, for example, make a case study
a type of news article in Whitehall or something like that. The
document's schema should not be tightly coupled to its groupings
in Whitehall's UI.