Add logic to re-render `choose_type` form when choosing 'group'
When choosing a 'group' option (such as `news_article`) then we
want to go back to the `choose_type` form to now display the
subtypes (e.g. `news_story`, `press_release`), and keep going back
to the form until we reach a 'leaf node' (in theory allowing for
multiple layers of depth).

Only when we reach a leaf node do we then redirect to the 'new
edition' form, with the correct configurable_document_type
pre-selected.