Include unpublished editions in revalidation jobs
We have a number of unpublished editions that remain live on GOV.UK because they were never successfully unpublished. This was generally because the unpublishing had an alternative URL which Whitehall considered valid at the time but Publishing API did not. We have subsequently tightened Whitehall's validation of the alternative URL attribute, but it would be useful to be able to identify the old invalid unpublishings so that we can potentially fix them.

Validating the unpublishing in the :publish validation context helps to make that clear to publishers so that they can understand why their unpublishing input may not have been reflected successfully on GOV.UK.