Add support for optionally updating document slug from the UI
Following the re-enabling of automatic slugging, this acts as a guardrail for users to decide when they want a document's slug to change. We have introduced a checkbox on the edition form for the user to decide if they want the slug changed.

Also, we now let the user know when the slug changes by updating the editorial remarks with a message showing the new slug.

Notes:
- The first draft of a document can always have its slug changed anytime. No editorial remark will be created.
- GDS Admin can still use the custom `edit_slug` route to update the slug of a published edition
- Updated `update_slug_if_possible` with a cleaner logic and moved `clear_slug` to [identifiable.rb](https://github.com/alphagov/whitehall/blob/edef6df12b64e96125afb719721380b3b3824e27/app/models/concerns/edition/identifiable.rb) in order to keep slug update things together in one place.

JIRA: https://gov-uk.atlassian.net/browse/WHIT-1889