Add 'groups' method for filtering types
This will be used in the next commit or two for surfacing 'groups'
in the 'choose type' (new document) screen.

Current behaviour:

- History pages
- News stories
- Press releases
- Government responses
- World news stories

Intended behaviour once we start making use of 'groups':

- History pages
- News articles

After choosing "News articles", the publisher would be taken to
an interstitial form where they can then specify the subtype of
news article. We don't want to surface all of the subtypes on
the first screen.