Add `configurable_document_group_ordering` property for ordering
We want to display the subtypes in order of usage - not
alphabetically. So we've added a `configurable_document_group_ordering`
property to do just that.

Before:

- Government response
- News story
- Press release
- World news story

After:

- News story
- Press release
- Government response
- World news story

HOWEVER: I think we'll soon want to come up with an additional
config file somewhere in Whitehall, to describe the _relationships_
between our configurable document types. It could describe which
content types are in the same 'group', which order they should appear
in within that group, and which order the group should appear in
relative to the other groups or top level types. It could also
power Whitehall search filtering and/or bulk republishing screens.
Not a blocker for this PR but worth considering soon.